requirements-1.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiohttp>=3.9.5
requests>=2.32.0
pymongo>=4.7.2
googletrans==4.0.0-rc1
pytz>=2024.1
Pillow>=10.3.0
colorama>=0.4.6
asyncio-throttle>=1.0.2
flask>=3.0.3
fastapi>=0.111.0
uvicorn>=0.30.1
cachetools>=5.3.3